Fixed missing type errors when compiling libraries due to the new variable type changes

+++ libpsn00b/include/psxgpu.h
@@ -615,7 +615,7 @@ DRAWENV *SetDefDrawEnv(DRAWENV *draw, int x, int y, int w, int h);
 // Debug font functions
 
 void FntLoad(int x, int y);
-char *FntSort(unsigned int *ot, char *pri, int x, int y, const char *text);
+char *FntSort(u_long *ot, char *pri, int x, int y, const char *text);
 int FntOpen(int x, int y, int w, int h, int isbg, int n);
 int FntPrint(int id, const char *fmt, ...);
 char *FntFlush(int id);
